requirements.txt
streamlit
scikit-learn
plotly
PyPDF2
python-docx